added database class file and rules file
