+++ utils.py
@@ -11,10 +11,26 @@ def between(fieldValue: str, bothCompareValues: str):
     return int(fieldValue) >= int(a) and int(fieldValue) <= int(b)
 
 
+
+
 evaluators = {
             ">": lambda x, y:int(x) > int(y),
             "<": lambda x, y: int(x) < int(y),
             "=": lambda x, y: int(x) == int(y),
             "!": lambda x, y: int(x) != int(y),
+            "#": lambda x, y: x in y,
+            "==": lambda x, y: x == y,
             "><": between
         }
+
+
+def valid_field(validation_string: str, value: str) -> bool:
+    """
+    Validate a value against a given condition.
+    """
+    operator, compare_value = validation_string.split(" ", 1)
+    evaluator = evaluators.get(operator)
+    if not evaluator:
+        raise ValueError(f"Invalid operator '{operator}'")
+        
+    return evaluator(value, compare_value)
